+++ src/gurobi_modelanalyzer/__init__.py
@@ -1,11 +1,12 @@
 __version__ = "v2.1.0"
 
+from .common import set_env
+
 from .results_analyzer import (
     kappa_explain,
     angle_explain,
     matrix_bitmap,
     converttofractions,
-    set_env,
 )
 
 from .solcheck import SolCheck
